requirements.txt
openai
openai-agents
modelcontext
mysql-connector-python
python-dotenv
